Convert to framework, add lexing, add unit tests
This was a doozy. Began by trying to get the original C code working from Apple's open source stuff. While I did get it working, I didn't want to deal with open source licensing and such, and it was very messy, so I spent a lot of time researching the HID spec to make this lexer. And it's Swifty too which is in the spirit of this project.

I like how the lexer came out. Pretty elegant with the enums I think.

Half way through I wanted to use unit tests to verify some bitwise parsing stuff. Apparently it's very difficult to get unit testing working for a shared library, so I ended up making a new framework project and replaced the .xcodeproj with it all. Unit tests are nice, but they'll need to be expanded.

Only problem with the lexed items so far is the Usage Page / Usage don't print the actual name, just the number. But this is fine, it'd need some sort of stack to properly do that as well as print indentation.