Add HIDCountryCode
I should really look at the specs more.

Also, I suppose it's possible for someone to manually make one of these enums (with a reserved & associated number) in an invalid way. They could init it with a value of 0, instead of using the .notSupported case. I suppose it's fine because getting the value will be fine, and SwiftyHID will never produce something invalid.